Shift remaining .sql.ts files into pg-meta
