+++ apps/studio/components/interfaces/Integrations/CronJobs/CronJobsTab.useCleanupActions.ts
@@ -4,8 +4,6 @@ import {
 } from '@supabase/pg-meta'
 import {
   CTID_BATCH_PAGE_SIZE,
-  getDeleteOldCronJobRunDetailsByCtidKey,
-  getJobRunDetailsPageCountKey,
   validatePageNumber,
 } from 'data/database-cron-jobs/database-cron-jobs.utils'
 import { useExecuteSqlMutation } from 'data/sql/execute-sql-mutation'
@@ -14,6 +12,10 @@ import { toast } from 'sonner'
 
 import { CLEANUP_INTERVALS } from './CronJobsTab.constants'
 import type { ConnectionVars } from '@/data/common.types'
+import {
+  getDeleteOldCronJobRunDetailsByCtidKey,
+  getJobRunDetailsPageCountKey,
+} from '@/data/database-cron-jobs/keys'
 import { useScheduleCronJobRunDetailsCleanupMutation } from '@/data/database-cron-jobs/schedule-clean-up-mutation'
 
 // Delay between batches to allow other queries to proceed (in milliseconds)
