+++ apps/studio/data/database-cron-jobs/keys.ts
@@ -1,3 +1,5 @@
+import { sqlKeys } from '../sql/keys'
+
 export const databaseCronJobsKeys = {
   create: () => ['cron-jobs', 'create'] as const,
   delete: () => ['cron-jobs', 'delete'] as const,
@@ -26,3 +28,17 @@ export const databaseCronJobsKeys = {
   ],
   timezone: (projectRef: string | undefined) => ['database-cron-timezone', projectRef] as const,
 }
+
+export const getJobRunDetailsPageCountKey = (projectRef: string | undefined) =>
+  sqlKeys.query(projectRef, ['cron-job-run-details', 'page-count'])
+
+export const getDeleteOldCronJobRunDetailsByCtidKey = (
+  projectRef: string | undefined,
+  interval: string,
+  startPage: number
+) => sqlKeys.query(projectRef, ['cron-job-run-details', 'delete-batch', interval, startPage])
+
+export const getScheduleDeleteCronJobRunDetailsKey = (
+  projectRef: string | undefined,
+  interval: string
+) => sqlKeys.query(projectRef, ['cron-job-run-details', 'schedule', interval])
