+++ apps/studio/data/database/keys.ts
@@ -1,3 +1,5 @@
+import { sqlKeys } from '../sql/keys'
+
 export const databaseKeys = {
   schemas: (projectRef: string | undefined) => ['projects', projectRef, 'schemas'] as const,
   keywords: (projectRef: string | undefined) => ['projects', projectRef, 'keywords'] as const,
@@ -49,3 +51,9 @@ export const databaseKeys = {
   supamonitorEnabled: (projectRef: string | undefined) =>
     ['projects', projectRef, 'supamonitor-enabled'] as const,
 }
+
+export const getLiveTupleEstimateKey = (
+  projectRef: string | undefined,
+  table: string,
+  schema = 'public'
+) => sqlKeys.query(projectRef, ['live-tuple-estimate', schema, table])
